pyflask/requirement.txt
Flask==2.0.1
simplejson==3.17.2
mysql-connector-python==8.0.25
